requirements.txt
accelerate==1.0.1
aiohappyeyeballs @ file:///croot/aiohappyeyeballs_1725434011349/work
aiohttp @ file:///croot/aiohttp_1725527756643/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1716290248287/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1728935693959/work
async-timeout @ file:///croot/async-timeout_1703096998144/work
attrs @ file:///croot/attrs_1729089401488/work
Bottleneck @ file:///croot/bottleneck_1707864210935/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
dataclasses-json==0.6.7
datasets @ file:///croot/datasets_1716911606380/work
dill @ file:///croot/dill_1715094664823/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
faiss==1.9.0
filelock @ file:///croot/filelock_1700591183607/work
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec @ file:///croot/fsspec_1714461537038/work
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1726922182064/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1727820890233/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1724778349782/work
huggingface_hub @ file:///croot/huggingface_hub_1724853938404/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
joblib==1.4.2
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///home/conda/feedstock_root/build_artifacts/jsonpointer_1725302897999/work
langchain @ file:///home/conda/feedstock_root/build_artifacts/langchain_1729752551769/work
langchain-community==0.3.3
langchain-core @ file:///home/conda/feedstock_root/build_artifacts/langchain-core_1729895987413/work
langchain-text-splitters @ file:///home/conda/feedstock_root/build_artifacts/langchain-text-splitters_1726389599337/work
langsmith @ file:///home/conda/feedstock_root/build_artifacts/langsmith_1727415227475/work
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
marshmallow==3.23.0
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
mpmath @ file:///croot/mpmath_1690848262763/work
multidict @ file:///croot/multidict_1701096859099/work
multiprocess @ file:///croot/multiprocess_1692294385131/work
mypy-extensions==1.0.0
networkx @ file:///croot/networkx_1720002482208/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
orjson @ file:///home/conda/feedstock_root/build_artifacts/orjson_1729643609037/work/target/wheels/orjson-3.10.10-cp310-cp310-linux_x86_64.whl#sha256=51dec98a64466469704b574dc83362e212a64ac41e271b8b97c936278d1f503f
packaging @ file:///croot/packaging_1720101850331/work
pandas @ file:///croot/pandas_1718308974269/work/dist/pandas-2.2.2-cp310-cp310-linux_x86_64.whl#sha256=a2434a313a5eb5b4fdbec6c770abfde859f643eff7b3a703791bf8b73e54c394
pillow==11.0.0
psutil==6.1.0
pyarrow @ file:///croot/pyarrow_1721664224170/work/python
pydantic @ file:///home/conda/feedstock_root/build_artifacts/pydantic_1726601062926/work
pydantic-settings==2.6.0
pydantic_core @ file:///home/conda/feedstock_root/build_artifacts/pydantic-core_1726524987213/work
PyMuPDF==1.24.12
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-dotenv==1.0.1
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex @ file:///croot/regex_1726669635272/work
requests @ file:///croot/requests_1721410876868/work
safetensors @ file:///croot/safetensors_1729663451734/work
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1708952932303/work
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1729066348927/work
sympy==1.13.1
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1720351771156/work
threadpoolctl==3.5.0
tokenizers @ file:///croot/tokenizers_1729696677247/work
torch==2.5.0
tqdm @ file:///croot/tqdm_1724853939799/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1729770468886/work
triton==3.1.0
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1727769808118/work
xxhash @ file:///croot/python-xxhash_1667919521615/work
yarl @ file:///croot/yarl_1725976495189/work
